// ui/src/services/cradle/models/AccessRequestNotification.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * CRADLE
 * Threat Intelligence Knowledge Management
 *
 * The version of the OpenAPI document: 2.10.2-beta.9bd46310 (2.10.2-beta.9bd46310)
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AccessRequestNotification
 */
export interface AccessRequestNotification {
    /**
     * 
     * @type {string}
     * @memberof AccessRequestNotification
     */
    readonly id?: string;
    /**
     * 
     * @type {string}
     * @memberof AccessRequestNotification
     */
    message: string;
    /**
     * 
     * @type {boolean}
     * @memberof AccessRequestNotification
     */
    isMarkedUnread?: boolean;
    /**
     * 
     * @type {number}
     * @memberof AccessRequestNotification
     */
    readonly entityId?: number;
    /**
     * 
     * @type {string}
     * @memberof AccessRequestNotification
     */
    readonly requestingUserId?: string;
    /**
     * 
     * @type {Date}
     * @memberof AccessRequestNotification
     */
    readonly timestamp?: Date;
    /**
     * 
     * @type {string}
     * @memberof AccessRequestNotification
     */
    readonly notificationType?: string;
}

/**
 * Check if a given object implements the AccessRequestNotification interface.
 */
export function instanceOfAccessRequestNotification(value: object): value is AccessRequestNotification {
    if (!('message' in value) || value['message'] === undefined) return false;
    return true;
}

export function AccessRequestNotificationFromJSON(json: any): AccessRequestNotification {
    return AccessRequestNotificationFromJSONTyped(json, false);
}

export function AccessRequestNotificationFromJSONTyped(json: any, ignoreDiscriminator: boolean): AccessRequestNotification {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'message': json['message'],
        'isMarkedUnread': json['is_marked_unread'] == null ? undefined : json['is_marked_unread'],
        'entityId': json['entity_id'] == null ? undefined : json['entity_id'],
        'requestingUserId': json['requesting_user_id'] == null ? undefined : json['requesting_user_id'],
        'timestamp': json['timestamp'] == null ? undefined : (new Date(json['timestamp'])),
        'notificationType': json['notification_type'] == null ? undefined : json['notification_type'],
    };
}

export function AccessRequestNotificationToJSON(json: any): AccessRequestNotification {
    return AccessRequestNotificationToJSONTyped(json, false);
}

export function AccessRequestNotificationToJSONTyped(value?: Omit<AccessRequestNotification, 'id'|'entity_id'|'requesting_user_id'|'timestamp'|'notification_type'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'message': value['message'],
        'is_marked_unread': value['isMarkedUnread'],
    };
}