// ui/src/services/cradle/models/NewUserNotification.ts
// @ts-nocheck
/* tslint:disable */
/* eslint-disable */
/**
 * CRADLE
 * Threat Intelligence Knowledge Management
 *
 * The version of the OpenAPI document: 2.10.2-beta.9bd46310 (2.10.2-beta.9bd46310)
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { EssentialUserRetrieve } from './EssentialUserRetrieve';
import {
    EssentialUserRetrieveFromJSON,
    EssentialUserRetrieveFromJSONTyped,
    EssentialUserRetrieveToJSON,
    EssentialUserRetrieveToJSONTyped,
} from './EssentialUserRetrieve';

/**
 * 
 * @export
 * @interface NewUserNotification
 */
export interface NewUserNotification {
    /**
     * 
     * @type {string}
     * @memberof NewUserNotification
     */
    readonly id?: string;
    /**
     * 
     * @type {string}
     * @memberof NewUserNotification
     */
    message: string;
    /**
     * 
     * @type {boolean}
     * @memberof NewUserNotification
     */
    isMarkedUnread?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof NewUserNotification
     */
    readonly timestamp?: Date;
    /**
     * 
     * @type {EssentialUserRetrieve}
     * @memberof NewUserNotification
     */
    newUser: EssentialUserRetrieve;
    /**
     * 
     * @type {string}
     * @memberof NewUserNotification
     */
    readonly notificationType?: string;
}

/**
 * Check if a given object implements the NewUserNotification interface.
 */
export function instanceOfNewUserNotification(value: object): value is NewUserNotification {
    if (!('message' in value) || value['message'] === undefined) return false;
    if (!('newUser' in value) || value['newUser'] === undefined) return false;
    return true;
}

export function NewUserNotificationFromJSON(json: any): NewUserNotification {
    return NewUserNotificationFromJSONTyped(json, false);
}

export function NewUserNotificationFromJSONTyped(json: any, ignoreDiscriminator: boolean): NewUserNotification {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'message': json['message'],
        'isMarkedUnread': json['is_marked_unread'] == null ? undefined : json['is_marked_unread'],
        'timestamp': json['timestamp'] == null ? undefined : (new Date(json['timestamp'])),
        'newUser': EssentialUserRetrieveFromJSON(json['new_user']),
        'notificationType': json['notification_type'] == null ? undefined : json['notification_type'],
    };
}

export function NewUserNotificationToJSON(json: any): NewUserNotification {
    return NewUserNotificationToJSONTyped(json, false);
}

export function NewUserNotificationToJSONTyped(value?: Omit<NewUserNotification, 'id'|'timestamp'|'notification_type'> | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'message': value['message'],
        'is_marked_unread': value['isMarkedUnread'],
        'new_user': EssentialUserRetrieveToJSON(value['newUser']),
    };
}